Change rapidsConf.shuffleCompressionLz4ChunkSize to a long

+++ sql-plugin/src/main/scala/com/nvidia/spark/rapids/TableCompressionCodec.scala
@@ -107,7 +107,7 @@ trait TableCompressionCodec {
 /**
  * A small case class used to carry codec-specific settings.
  */
-case class TableCompressionCodecConfig(lz4ChunkSize: Int)
+case class TableCompressionCodecConfig(lz4ChunkSize: Long)
 
 object TableCompressionCodec {
   private val codecNameToId = Map(
